Read description
Added controller implementation. If testing with controller set
testingWithKeyboard bool to false in PlayerControllerScript.

Added a gun that rotates.

Cleaned up code. Try to keep as many things private as possible because
if you make floats and such public whatever is in the inspector goes
and it could make debugging hard.

If curious about how InControl works read stuff in the InControl folder.